Add Observable and Observer classes to utils for observer pattern implementation
